+++ src/redux/actions/goalsActions.ts
@@ -1,13 +1,16 @@
 import { createAction, createAsyncThunk } from '@reduxjs/toolkit';
 import GoalsService from '~http/services/goals';
+import { triggerReloadStat } from '~redux/actions/statisticActions';
+import { AppThunkAPI } from '~redux/store/configureStore';
 
 const PREFIX = 'GOALS';
 
 export const setGoal = createAction<number>(`${PREFIX}/setGoal`);
 
-export const deleteUserGoalItem = createAsyncThunk(`${PREFIX}/deleteUserGoalItem`, async (id: string) => {
+export const deleteUserGoalItem = createAsyncThunk(`${PREFIX}/deleteUserGoalItem`, async (id: string, { dispatch }: AppThunkAPI) => {
   try {
     const { data } = await GoalsService().deleteUserGoalItem({ id });
+    dispatch(triggerReloadStat());
     return data;
   } catch (error) {
     console.error(error);
@@ -47,10 +50,11 @@ export const getGoalItems = createAsyncThunk(`${PREFIX}/getGoalItems`, async () 
   }
 });
 
-export const addGoalItem = createAsyncThunk(`${PREFIX}/addGoalItem`, async (pages: string | null) => {
+export const addGoalItem = createAsyncThunk(`${PREFIX}/addGoalItem`, async (pages: string | null, { dispatch }: AppThunkAPI) => {
   const params = { pages, added_at: new Date().getTime() };
   try {
     const { data } = await GoalsService().addGoalItem({ ...params });
+    dispatch(triggerReloadStat());
     return data;
   } catch (error) {
     console.error(error);