+++ src/screens/Goals/GoalDetails/styles.ts
@@ -186,4 +186,7 @@ export default StyleSheet.create({
     fontSize: 16,
     color: colors.neutral_light,
   },
+  listFooterComponent: {
+    height: 80,
+  },
 });